fix: removinfg snippet viewer and additional changes
